First batch of rewrites
I'm rewriting a lot of the library due to some changes @Niki4tap suggested.

These changes mainly include the way features are handled, and the way file formats are handled.

After this rewrite users will be able to make config files of different file formats in the same project, while also keeping the speed benefits of code being removed when the features for the file formats are disabled.