Taking a break now, final changes
Kept trying to fix docs.rs for the past hour, so apologies for the multiple "taking a break this is the last commit" type messages xD
This should be the last one as I've managed to finally figure out what the issue was.

I've also made some smaller changes relating to updating the license on crates.io and the maintenance status.